% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfkBR.R
\name{fun_cal_gfk_br}
\alias{fun_cal_gfk_br}
\alias{gfkBR}
\title{Cálculo do índice GFK BR}
\usage{
gfkBR()
}
\arguments{
\item{ipca_ponderado}{Corresponde a base de dados do IPCA limpa e organizada com \code{date}, \code{product_groups}, \code{items}, \code{ipca_mensal}, \code{ipca_peso}.}

\item{db_gfk}{Corresponde aos dados para o cáculo do índice GFK. Primeira aba com \code{date}, \code{uf}, \code{product_groups}, \code{valor}; segunda aba com \code{date}, \code{product_groups}, \code{valor}.}
}
\value{
Retorna um data.frame com o índice GFK BR calculado
\code{gfk_br} - corresponde ao índice GFK agregado calculado
}
\description{
Função para calcular o índice GFK BR, isto é, em nível agregado.
}
\details{
O arquivo de input deve corresponder a base de dados do IPCA e do índice GFK limpas e organizadas.

\code{Mês} - coluna de datas no formato MM/YYYY;
\code{Variável} - coluna corresponde a variação do IPCA e do peso dos items que compõe o IPCA;
\code{Geral, grupo, subgrupo, item e subitem} - corresponde ao grupo de produtos do item;
\code{valor} - valor corresponde a variação nominal do IPCA e ao peso dos items que compõe o IPCA.

\code{date} - coluna de datas no formato YYYY/MM;
\code{uf} - coluna das unidades federativas;
\code{product_groups} - nome do grupo de produtos classificado;
\code{valor} - valor nominal.
}
\examples{
\dontrun{
indice_gfk = fun_cal_gfk_br()
}

}
\author{
Luiz Paulo
}
