% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcaPonderado.R
\name{fun_cal_IPCA}
\alias{fun_cal_IPCA}
\alias{ipcaPonderado}
\title{Cálculo do IPCA ponderado pelos pesos dos items}
\usage{
ipcaPonderado()
}
\arguments{
\item{db_ipca}{Corresponde a base de dados do IPCA limpa e organizada com \code{date}, \code{product_groups}, \code{items}, \code{ipca_mensal}, \code{ipca_peso}.}
}
\value{
Retorna um data.frame com o IPCA ponderado e com o índice do IPCA ponderado.
\code{ipca_group} - corresponde ao ipca ponderado e ao índice do IPCA.
}
\description{
Função para calcular o IPCA ponderado pelos pesos dos items que compõe o índice GFK.
}
\details{
O arquivo de input deve corresponder a base de dados do IPCA limpa e organizada.

\code{date} - coluna de datas no formato YYYY/MM/DD;
\code{items} - corresponde aos items que compõe GFK;
\code{product_groups} - corresponde ao grupo de items;
\code{ipca_mensal} - corresponde a variação nominal do IPCA;
\code{ipca_peso} - corresponde ao peso atribuido aos items do IPCA.
}
\examples{
\dontrun{
ipca_ponderado = fun_cal_IPCA()
}

}
\author{
Luiz Paulo
}
