% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessaz.R
\name{fun_dessaz_gfk}
\alias{fun_dessaz_gfk}
\alias{dessaz_gfk}
\title{Calculando o ajuste sazonal do índice GFK BR & GFK UF}
\usage{
dessaz_gfk(arima.model = "")
}
\arguments{
\item{indice_gfk_uf}{corresponde ao índice GFK UF calculado \code{gfk_uf};}

\item{indice_gfk}{corresponde ao índice GFK em nível nacional \code{gfk_br};}

\item{data_calendar}{corresponder ao calenário para ajuste sazonal \code{peso_carnval}, \code{peso_corpus_christi}.}
}
\value{
Retorna um data.frame com o índice GFK BR-UF com ajuste sazonal.
\code{db_gfk_dessaz} - corresponde ao índice GFK BR-UF com ajuste sazonal.
}
\description{
Função para calcular o ajuste sazonal do índice GFK UF e em nível Nacional
}
\details{
O arquivo de input deve corresponder ao índice GFK BR-UF e ao calendário.

\code{date} - coluna de datas no formato YYYY/MM/DD;
\code{gfk_uf} - coluna corresponde ao índice GFK UF calculado;
\code{gfk_br} - corresponde ao índice GFK BR calculado;
\code{data_calendar} - corresponde ao calendário para o ajuste sazonal.
}
\examples{
\dontrun{
bd_gfk <- fun_dessaz_gfk(arima.model = "(0 1 1) (0 1 1)")
}

}
\author{
Luiz Paulo
}
