% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{fun_importCalendar}
\alias{fun_importCalendar}
\alias{importCalendar}
\title{funcão calendário}
\usage{
importCalendar()
}
\value{
O retorno é um objeto ts (time series) com os dados do calendário
\code{calendar_ts} - dados, em objeto ts, do calendário.
}
\description{
Função para importar e organizar a base de dados do calendário para o ajuste sazonal
}
\details{
O arquivo de input deve corresponder a base de dados calendário:

\code{date} - coluna de datas no formato YYYY/MM;
\code{peso_carnaval} - coluna com o calendário do carnaval;
\code{peso_corpus christi} - coluna com o calendário do Corpus Christi.
}
\examples{
\dontrun{
data_calendar = fun_importCalendar()
}

}
\author{
Luiz Paulo
}
