% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importClean.R
\name{fun_importCleanGFK}
\alias{fun_importCleanGFK}
\alias{importClean}
\title{Importação e Limpeza do Banco de dados GFK}
\usage{
importClean(path_raw)
}
\arguments{
\item{DB_GFK.xlsx}{Um arquivo em formato xlsx, com duas abas, contendo os dados para o cáculo do índice GFK. Primeira aba com \code{date}, \code{uf}, \code{product_groups}, \code{valor}; segunda aba com \code{date}, \code{product_groups}, \code{valor}.}
}
\value{
O retorno é uma lista com o banco de dados limpo e organizado para as unidades federativas e em nível nacional (agregado);
\code{db_region} - lista com os dados das unidades federativas;
\code{db_br} - lista com os dados agregados em nível nacional.
}
\description{
Função para importar, limpar e organizar a base de dados do índice GFK
}
\details{
O arquivo de input deve corresponder a base de dados que compõe o índice GFK.
Cada linha representa um mês e tem informações como:

\code{date} - coluna de datas no formato YYYY/MM;
\code{uf} - coluna das unidades federativas;
\code{product_groups} - nome do grupo de produtos classificado;
\code{valor} - valor nominal.
}
\examples{
\dontrun{
db_gfk = fun_importCleanGFK(path_raw = "DB_GFK.xlsx")
}

}
\author{
Luiz Paulo
}
