% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated.R
\name{fun_updated()}
\alias{fun_updated()}
\alias{updated}
\title{Atualização do Banco de dados GFK}
\usage{
updated()
}
\value{
O retorno é uma lista com o banco de dados atualizada
\code{db_updated_region} - lista com os dados das unidades federativas;
\code{db_updated_br} - lista com os dados agregados em nível nacional.
}
\description{
Função para atualizar a base de dados do índice GFK
}
\details{
O arquivo de input deve corresponder a base de dados que compõe o índice GFK.
Cada linha representa um mês e tem informações como:

\code{date} - coluna de datas no formato YYYY/MM;
\code{uf} - coluna das unidades federativas;
\code{product_groups} - nome do grupo de produtos classificado;
\code{valor} - valor nominal.
}
\examples{
\dontrun{
db_gfk = fun_updated()
}

}
\author{
Luiz Paulo
}
