% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfk.R
\name{func_get_gfk}
\alias{func_get_gfk}
\alias{get_gfk}
\title{Função pipe com todas as funções acopladas para cálculo do índice GFK}
\usage{
get_gfk(dir_gfk = as.character())
}
\description{
Função encapsula todas as funções utilizadas no índice GFK
}
\details{
O arquivo de input deve corresponder a todas as funções
}
\examples{
\dontrun{
db_gfk = get_gfk()
}

}
\author{
Luiz Paulo
}
